--- client/src/pages/Libro.jsx
@@ -1,11 +1,0 @@
-import Header from '../components/Header';
-import Actions from '../components/Actions';
-import Preview from '../components/Preview';
-
-function Libro() {
-  return (
-    <Preview />
-  );
-}
-
-export default Libro;
